# docs/swagger.yaml
basePath: /task
definitions:
  types.PostObjectHandlerRequest:
    properties:
      compiler:
        example: python3
        type: string
      task:
        example: your code
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: This is a HTTP-server for code processing.
  title: CodeProcessor
  version: "1.0"
paths:
  /:
    post:
      consumes:
      - application/json
      description: |-
        Создание новой задачи, старт работы обработчика.
        Возвращает ID задачи
      parameters:
      - description: Task and language name
        in: body
        name: Object
        required: true
        schema:
          $ref: '#/definitions/types.PostObjectHandlerRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Код успешно загружен
          schema:
            type: string
        "400":
          description: Bad request
          schema:
            type: string
      summary: Отправка кода и названия языка программирования
      tags:
      - object
  /result:
    get:
      consumes:
      - application/json
      description: Возвращает результат задачи с указанным ID
      parameters:
      - description: task_id
        in: query
        name: ID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad request
          schema:
            type: string
      summary: Получение результата
      tags:
      - object
  /status:
    get:
      consumes:
      - application/json
      description: |-
        Проверяет статус выполнения задачи с указанным ID
        Возвращает статус выполнения задачи с указанным ID
      parameters:
      - description: task_id
        in: query
        name: ID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad request
          schema:
            type: string
      summary: Проверка статуса выполнения
      tags:
      - object
swagger: "2.0"